+++ packages/dso-toolkit/src/components/hero-image/index.ts
@@ -1,0 +1,2 @@
+export * from "./hero-image.models.js";
+export * from "./hero-image.stories-of.js";