+++ storybook/src/components/hero-image/hero-image.content.ts
@@ -1,0 +1,1 @@
+export const image = "images/banner-image.webp";