+++ storybook/src/components/hero-image/hero-image.core-stories.ts
@@ -1,0 +1,28 @@
+import readme from "@dso-toolkit/core/src/components/hero-image/readme.md?raw";
+import type { Meta } from "@storybook/web-components-vite";
+import { heroImageMeta, heroImageStories } from "dso-toolkit";
+
+import { templateContainer } from "../../templates";
+
+import { image } from "./hero-image.content";
+
+const meta: Meta = {
+  ...heroImageMeta({ readme }),
+  title: "Core/Hero Image",
+};
+
+export default meta;
+
+const { Default } = heroImageStories({
+  templateContainer,
+  storyTemplates: (templates) => {
+    const { heroImageTemplate } = templates;
+
+    return {
+      heroImageTemplate,
+      image,
+    };
+  },
+});
+
+export { Default };